requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
alembic==1.13.3
annotated-types==0.7.0
APScheduler==3.10.4
asyncpg==0.29.0
attrs==24.2.0
certifi==2024.8.30
frozenlist==1.4.1
greenlet==3.1.1
idna==3.10
magic-filter==1.0.12
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.1.0
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
pydentic==0.0.1.dev3
python-dotenv==1.0.1
python-stdnum==1.20
pytz==2024.2
six==1.16.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
tzlocal==5.2
yarl==1.13.1
